dom/base/test/test_bug1576154.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1576154
-->
<head>
  <title>Test for Bug 1576154</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1576154">Mozilla Bug 1576154</a>
<p id="display"></p>
<!-- bug1576154.sjs returns an SVG image with HTTP error code 500. -->
<img src="bug1576154.sjs">
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 657191 **/
SimpleTest.waitForExplicitFinish();

window.onload = function() {
  let img = document.querySelector("img");
  img.src = "green.png";
  img.onload = function() {
    // As long as this doesn't crash, this test passes.
    ok(true, "test passed");
    SimpleTest.finish();
  };
}

</script>
</pre>
</body>
</html>
